@@ -1,7 +1,7 @@
 import { LOCALE_ID, NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { CookieService } from 'ngx-cookie-service';
@@ -14,23 +14,17 @@ import { ReactiveFormsModule } from '@angular/forms';
 
 
 registerLocaleData(localeEs, 'es');
-@NgModule({
-  declarations: [
-    AppComponent,
-    LoginComponent,
-  ],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    HttpClientModule,
-    ReactiveFormsModule,
-    BrowserAnimationsModule,
-    NgxPrintModule,
-  ],
-  providers: [
-    CookieService,
-    {provide: LOCALE_ID, useValue: 'es'},
-  ],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        LoginComponent,
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        AppRoutingModule,
+        ReactiveFormsModule,
+        BrowserAnimationsModule,
+        NgxPrintModule], providers: [
+        CookieService,
+        { provide: LOCALE_ID, useValue: 'es' },
+        provideHttpClient(withInterceptorsFromDi()),
+    ] })
 export class AppModule { }
